requirements.txt
aiogram~=3.4
aiohttp~=3.9
aiosqlite~=0.19
fastapi~=0.109
uvicorn~=0.27
python-dotenv~=1.0
